refactor: cache 모듈 파일/폴더명 수정 (#49)
