--- src/modules/apiCache/index.ts
@@ -1,3 +1,0 @@
-export * from './class/ApiCache';
-export * from './utils/apiCache';
-export * from './constants/apiCache';
